# Stock-Dashboard/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Warehouse Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /tool-requests:
    get:
      operationId: listToolRequests
      tags: [tool-requests]
      summary: List tool borrowing requests
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: requesterId
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: List of tool requests
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ToolRequest"
    post:
      operationId: createToolRequest
      tags: [tool-requests]
      summary: Create a new tool borrowing request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ToolRequestInput"
      responses:
        "201":
          description: Created tool request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ToolRequest"
  /tool-requests/{id}:
    get:
      operationId: getToolRequest
      tags: [tool-requests]
      summary: Get a tool request by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Tool request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ToolRequest"
    delete:
      operationId: deleteToolRequest
      tags: [tool-requests]
      summary: Delete a tool request
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /tool-requests/{id}/approve:
    post:
      operationId: approveToolRequest
      tags: [tool-requests]
      summary: Approve a tool request
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApprovalInput"
      responses:
        "200":
          description: Updated request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ToolRequest"
  /tool-requests/{id}/return:
    post:
      operationId: returnToolRequest
      tags: [tool-requests]
      summary: Mark tool as returned
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ToolReturnInput"
      responses:
        "200":
          description: Updated request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ToolRequest"

  # ── STOCK MOVEMENTS ──────────────────────────────────────────────────────────
...
  /suppliers:
    get:
      operationId: listSuppliers
      tags: [suppliers]
      summary: List suppliers
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: category
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of suppliers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Supplier"
    post:
      operationId: createSupplier
      tags: [suppliers]
      summary: Create a supplier
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SupplierInput"
      responses:
        "201":
          description: Created supplier
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Supplier"
  /suppliers/{id}:
    get:
      operationId: getSupplier
      tags: [suppliers]
      summary: Get supplier by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Supplier
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Supplier"
    patch:
      operationId: updateSupplier
      tags: [suppliers]
      summary: Update supplier
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SupplierInput"
      responses:
        "200":
          description: Updated supplier
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Supplier"
    delete:
      operationId: deleteSupplier
      tags: [suppliers]
      summary: Delete supplier
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── PURCHASE ORDERS ────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    ToolRequest:
      type: object
      required:
        - id
        - requestNumber
        - requesterName
        - department
        - status
        - toolId
        - toolName
        - scheduledStart
        - scheduledEnd
        - createdAt
      properties:
        id:
          type: integer
        requestNumber:
          type: string
        requesterName:
          type: string
        requesterId:
          type: ["integer", "null"]
        department:
          type: string
        toolId:
          type: integer
        toolName:
          type: string
        toolCode:
          type: string
        status:
          type: string
          enum: [pending, approved, rejected, in_use, returned, overdue]
        scheduledStart:
          type: string
        scheduledEnd:
          type: string
        actualReturn:
          type: ["string", "null"]
        purpose:
          type: ["string", "null"]
        conditionOnReturn:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        approvedBy:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
    ToolRequestInput:
      type: object
      required: [requesterName, department, toolId, scheduledStart, scheduledEnd]
      properties:
        requesterName:
          type: string
        requesterId:
          type: integer
        department:
          type: string
        toolId:
          type: integer
        scheduledStart:
          type: string
        scheduledEnd:
          type: string
        purpose:
          type: string
        notes:
          type: string
    ToolReturnInput:
      type: object
      required: [conditionOnReturn]
      properties:
        conditionOnReturn:
          type: string
          enum: [good, fair, poor, damaged]
        notes:
          type: string

    # ── Approval ───────────────────────────────────────────────────────────────
    ApprovalInput:
      type: object
      required: [approvedBy]
      properties:
        approvedBy:
          type: string
        notes:
          type: string

    # ── Stock Movement ─────────────────────────────────────────────────────────
...
    Supplier:
      type: object
      required: [id, code, name, category, isActive, createdAt]
      properties:
        id:
          type: integer
        code:
          type: string
        name:
          type: string
        category:
          type: string
        contactPerson:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        city:
          type: ["string", "null"]
        npwp:
          type: ["string", "null"]
        paymentTerms:
          type: ["string", "null"]
        leadTimeDays:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        isActive:
          type: boolean
        createdAt:
          type: string
        updatedAt:
          type: string
    SupplierInput:
      type: object
      required: [name, category]
      properties:
        name:
          type: string
        category:
          type: string
        contactPerson:
          type: string
        email:
          type: string
        phone:
          type: string
        address:
          type: string
        city:
          type: string
        npwp:
          type: string
        paymentTerms:
          type: string
        leadTimeDays:
          type: string
        notes:
          type: string
        isActive:
          type: boolean

    # ── Purchase Order ──────────────────────────────────────────────────────────
...