# Stock-Dashboard/lib/api-zod/src/generated/types/toolRequest.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Warehouse Management System API
 * OpenAPI spec version: 0.1.0
 */
import type { ToolRequestStatus } from './toolRequestStatus';

export interface ToolRequest {
  id: number;
  requestNumber: string;
  requesterName: string;
  /** @nullable */
  requesterId?: number | null;
  department: string;
  toolId: number;
  toolName: string;
  toolCode?: string;
  status: ToolRequestStatus;
  scheduledStart: string;
  scheduledEnd: string;
  /** @nullable */
  actualReturn?: string | null;
  /** @nullable */
  purpose?: string | null;
  /** @nullable */
  conditionOnReturn?: string | null;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  approvedBy?: string | null;
  createdAt: string;
  updatedAt?: string;
}